move snapped_window to utils
